FEAT: Update color of each component
