+++ src/components/modules/PlaceMarker.tsx
@@ -32,7 +32,9 @@ const PlaceMarker: React.FC<Props> = ({
               // Set (width * scale / 2, height * scale / 2) to move center of icon to position.
               anchor: new google.maps.Point(10, 10),
               path: 'M20 10C20 15.5228 15.5228 20 10 20C4.47715 20 0 15.5228 0 10C0 4.47715 4.47715 0 10 0C15.5228 0 20 4.47715 20 10Z',
-              fillColor: selected ? theme.palette.primary.light : '#D95140',
+              fillColor: selected
+                ? theme.palette.secondary.light
+                : theme.palette.primary.dark,
               fillOpacity: 1,
               scale: 1,
               strokeColor: '#FFF',
